#171: Move BatchedDataStream to CsvUtils. Fix typing when reading CSV
